# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: FerreMax CRM/ERP API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Unauthorized
...
  /customers:
    get:
      operationId: listCustomers
      tags: [customers]
      summary: List customers
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: type
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
        - name: offset
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Customers list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Customer"
    post:
      operationId: createCustomer
      tags: [customers]
      summary: Create customer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CustomerInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
  /customers/{id}:
    get:
      operationId: getCustomer
      tags: [customers]
      summary: Get customer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Customer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
    patch:
      operationId: updateCustomer
      tags: [customers]
      summary: Update customer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CustomerUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
    delete:
      operationId: deleteCustomer
      tags: [customers]
      summary: Delete customer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /customers/{id}/purchases:
    get:
      operationId: getCustomerPurchases
      tags: [customers]
      summary: Get customer purchase history
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Purchase history
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Sale"

  # ─── CATEGORIES ──────────────────────────────────────────────────────────────
...
  /ai/config:
    get:
      operationId: getAiConfig
      tags: [ai]
      summary: Get AI configuration
      responses:
        "200":
          description: AI config
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiConfig"
    patch:
      operationId: updateAiConfig
      tags: [ai]
      summary: Update AI configuration
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AiConfigUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiConfig"
  /ai/chat:
    post:
      operationId: aiChat
      tags: [ai]
      summary: Chat with AI assistant
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AiMessage"
      responses:
        "200":
          description: AI response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiReply"

  # ─── KNOWLEDGE BASE ──────────────────────────────────────────────────────────
  /knowledge:
    get:
      operationId: listKnowledgeItems
      tags: [knowledge]
      summary: List knowledge base items
      parameters:
        - name: category
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Knowledge items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/KnowledgeItem"
    post:
      operationId: createKnowledgeItem
      tags: [knowledge]
      summary: Create knowledge item
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/KnowledgeInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KnowledgeItem"
  /knowledge/{id}:
    patch:
      operationId: updateKnowledgeItem
      tags: [knowledge]
      summary: Update knowledge item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/KnowledgeUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KnowledgeItem"
    delete:
      operationId: deleteKnowledgeItem
      tags: [knowledge]
      summary: Delete knowledge item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── USERS ───────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string
    AuthResponse:
      type: object
      required: [token, user]
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"

    # ── DASHBOARD ─────────────────────────────────────────────────────────────
...
    Customer:
      type: object
      required: [id, name, phone, type, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        phone:
          type: string
        whatsapp:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        city:
          type: ["string", "null"]
        type:
          type: string
        document:
          type: ["string", "null"]
        company:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        tags:
          type: ["string", "null"]
        totalPurchases:
          type: ["number", "null"]
        createdAt:
          type: string
    CustomerInput:
      type: object
      required: [name, phone, type]
      properties:
        name:
          type: string
        phone:
          type: string
        whatsapp:
          type: string
        email:
          type: string
        address:
          type: string
        city:
          type: string
        type:
          type: string
        document:
          type: string
        company:
          type: string
        notes:
          type: string
        tags:
          type: string
    CustomerUpdate:
      type: object
      properties:
        name:
          type: string
        phone:
          type: string
        whatsapp:
          type: string
        email:
          type: string
        address:
          type: string
        city:
          type: string
        type:
          type: string
        document:
          type: string
        company:
          type: string
        notes:
          type: string
        tags:
          type: string

    # ── CATEGORY ──────────────────────────────────────────────────────────────
...
    SaleItem:
      type: object
      required: [productId, productName, quantity, unitPrice, total]
      properties:
        productId:
          type: integer
        productName:
          type: string
        quantity:
          type: integer
        unitPrice:
          type: number
        discount:
          type: ["number", "null"]
        total:
          type: number
    Sale:
      type: object
      required: [id, total, status, paymentMethod, createdAt]
      properties:
        id:
          type: integer
        customerId:
          type: ["integer", "null"]
        customerName:
          type: ["string", "null"]
        items:
          type: array
          items:
            $ref: "#/components/schemas/SaleItem"
        subtotal:
          type: number
        discount:
          type: ["number", "null"]
        tax:
          type: ["number", "null"]
        total:
          type: number
        status:
          type: string
        paymentMethod:
          type: string
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
...
    AiConfig:
      type: object
      required: [id, model, temperature, maxTokens, enabled]
      properties:
        id:
          type: integer
        apiKey:
          type: ["string", "null"]
        model:
          type: string
        temperature:
          type: number
        maxTokens:
          type: integer
        systemPrompt:
          type: ["string", "null"]
        enabled:
          type: boolean
        autoReplyEnabled:
          type: boolean
        businessHoursStart:
          type: ["string", "null"]
        businessHoursEnd:
          type: ["string", "null"]
    AiConfigUpdate:
      type: object
      properties:
        apiKey:
          type: string
        model:
          type: string
        temperature:
          type: number
        maxTokens:
          type: integer
        systemPrompt:
          type: string
        enabled:
          type: boolean
        autoReplyEnabled:
          type: boolean
        businessHoursStart:
          type: string
        businessHoursEnd:
          type: string
    AiMessage:
      type: object
      required: [message]
      properties:
        message:
          type: string
        conversationId:
          type: integer
    AiReply:
      type: object
      required: [response]
      properties:
        response:
          type: string

    # ── KNOWLEDGE ─────────────────────────────────────────────────────────────
    KnowledgeItem:
      type: object
      required: [id, title, content, category, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        content:
          type: string
        category:
          type: string
        createdAt:
          type: string
    KnowledgeInput:
      type: object
      required: [title, content, category]
      properties:
        title:
          type: string
        content:
          type: string
        category:
          type: string
    KnowledgeUpdate:
      type: object
      properties:
        title:
          type: string
        content:
          type: string
        category:
          type: string

    # ── USERS ─────────────────────────────────────────────────────────────────
    User:
      type: object
      required: [id, name, email, role, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
        active:
          type: boolean
        createdAt:
          type: string
...